+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasetop.py
@@ -87,7 +87,7 @@ class RetrieveDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-08-27T08:20:42.765Z")
+    updated: Optional[datetime] = parse_datetime("2026-08-27T13:21:49.218Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")